requirements.txt
django==2.2
djangorestframework==3.9.2
celery==4.4.7
redis==3.2.0
django_celery_results
celery-progress
pyVmomi
Pillow
"celery[redis]"
